refactor(modularity): Phase 1d - eliminate orchestra + agents cross-module re-exports
Remove all cross-module re-exports from orchestra/__init__.py (18 symbols)
and agents/__init__.py (2 symbols). Redirect affected consumer files to
import from source modules via their public API (__all__).

**Orchestra changes:**
- Remove eager imports: GitClient, OrchestraConfig
- Remove lazy imports for domain, observability, utils, services, models symbols
- Keep all self-module re-exports from vibe3.orchestra.* submodules
- Update module docstring to clarify self-module re-exports only

**Agents changes:**
- Remove eager import: PromptContextMode (from models)
- Remove lazy import: build_tools_guide_section (from prompts.sections)
- Update docstring and README to reflect correct dependencies

**Consumer file updates:**
- orchestra/failed_gate.py: Import GateResult, GateStatus from vibe3.domain
- orchestra/dispatch_health_check.py: Import append_orchestra_event from vibe3.observability
- orchestra/global_dispatch_coordinator.py: Import MAX_INTENTS_PER_TICK, QueueEntry from source modules
- orchestra/queue_operations.py: Split imports between self-module and cross-module sources
- server/registry.py: Import orchestra_events_log_path, orchestra_log_dir from vibe3.observability
- server/app.py: Import orchestra_events_log_path, orchestra_log_dir, setup_logging from vibe3.observability
- commands/status.py: Import orchestra_events_log_path from vibe3.observability
- commands/scan.py: Import append_governance_event from vibe3.observability

**Test fixes:**
- Update test_failed_gate.py to mock vibe3.observability.setup_logging

This completes Phase 1 of the modularity refactoring: all top-level modules
now import from their own submodules or from other modules' public APIs,
eliminating cross-module re-export chains.

Related: #2281 (Phase 1c), #2282 (this phase)
