+++ src/vibe3/commands/status.py
@@ -22,7 +22,7 @@ from vibe3.commands.common import (
 )
 from vibe3.config.env_override import OVERRIDE_RULES
 from vibe3.models import OrchestraConfig
-from vibe3.orchestra import orchestra_events_log_path
+from vibe3.observability import orchestra_events_log_path
 from vibe3.server import validate_pid_file
 from vibe3.services import get_manager_usernames
 from vibe3.ui import console
