+++ src/vibe3/orchestra/queue_operations.py
@@ -5,15 +5,10 @@ from __future__ import annotations
 import time
 from typing import TYPE_CHECKING, Callable
 
+from vibe3.domain import QualifyGateService, find_role_for_state
 from vibe3.models import IssueInfo, IssueState, OrchestraConfig, QueueEntry
-from vibe3.orchestra import (
-    QualifyGateService,
-    append_orchestra_event,
-    find_role_for_state,
-    get_flow_context,
-    is_auto_task_branch,
-    load_issue,
-)
+from vibe3.observability import append_orchestra_event
+from vibe3.orchestra import get_flow_context, is_auto_task_branch, load_issue
 from vibe3.utils.queue_ordering import sort_ready_issues
 
 if TYPE_CHECKING:
